TF Docs adds extra files on git_add when using working directory '.'
When using TF Docs after other steps which contain lint actions for example which output additional files to the current working directory TF Docs will add those files. With the following configuration at the end of a workflow run I get additional files in my PRs. I believe this is because the `OUTPUT_FILES` variable is not defined. It appears it should be $INPUT_OUTPUT_FILE.

To replicate, create a job which `touch`s some files and then runs this step.

```
    - name: TF-Docs
      uses: terraform-docs/gh-actions@v1.0.0
      with:
        working-dir: .
        output-file: README.md
        output-method: inject
        git-push: "true"
```

Signed-off-by: Jack Penson <penson.122@gmail.com>